# docs/api/commerceflow-openapi.yaml
...
openapi: 3.0.3
info:
  title: CommerceFlow API
  version: 1.0.0
  description: |
    Microservices-based e-commerce order processing.

    Placing an order is **asynchronous**. `POST /api/orders` returns `201` with status `CREATED`
    as soon as the order is accepted and the orchestrator has been started; stock
    reservation, payment and notification then run over Kafka. Poll `GET /api/orders/{id}` to
    follow it — the status walks
    `CREATED → INVENTORY_RESERVED → PAID → COMPLETED`, or moves to `CANCELLED` with
    `failedStep` and `failureReason` telling you where and why it stopped.

    Every response is wrapped in a uniform envelope: `ApiResponse` on success, `ErrorResponse`
    on failure. Both carry the `correlationId`, which is the id to quote in a support request —
    it stitches the whole saga together in the logs.
  contact:
    name: CommerceFlow Platform Team
    email: platform@commerceflow.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local API Gateway (docker compose)
  - url: https://api.commerceflow.io
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /api/orders:
    post:
      tags: [Orders]
      summary: Place an order
      description: |
        Persists the order and starts the saga in one transaction, then returns
        immediately with status `CREATED`. Stock reservation and payment happen asynchronously.

        Supplying `idempotencyKey` makes a retried request return the original order instead of
        placing a second one.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderRequest" }
      responses:
        "201":
          description: Order accepted; the saga has started
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: A product in the basket does not exist (`PRODUCT_NOT_FOUND`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: A product is no longer for sale, or the basket mixes currencies
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: The catalogue is unavailable, so the order cannot be priced
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [Orders]
      summary: List orders
      description: A customer sees their own orders; an `ADMIN` sees every order.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [CREATED, INVENTORY_RESERVED, PAID, COMPLETED, CANCELLED]
        - name: userId
          in: query
          schema: { type: string, format: uuid }
          description: ADMIN only; ignored for customers
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, totalAmount, status, orderNumber]
            default: createdAt
        - { name: direction, in: query, schema: { type: string, enum: [asc, desc], default: desc } }
      responses:
        "200":
          description: A page of orders
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/Page"
                          - type: object
                            properties:
                              content:
                                type: array
                                items: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Fetch one order
      description: |
        Served from the CQRS read model. A customer may only read their own orders; an unknown
        order and a foreign order both return `404`, so the endpoint cannot be used to discover
        which order ids exist.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The order
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel an order
      description: |
        Cancels the order and unwinds whatever it had already done: a refund if it was paid for,
        and the stock back either way.

        **The order is marked `CANCELLED` immediately**, before any of the undoing. It will not be
        fulfilled, and that is knowable right now. Whether the money has come back is the
        *payment's* status rather than the order's — there is no `REFUNDED` order status — so poll
        `GET /api/payments/order/{orderId}` for that; it moves to `REFUNDED` when the money is on
        its way back.

        A customer may cancel their own order; an administrator may cancel anyone's. A foreign
        order returns `404` rather than `403`, for the same reason as `GET /api/orders/{id}`.

        ### Why this can be refused

        | Situation | Status | What the caller should do |
        |---|---|---|
        | Saga still running | `409` | Try again in a few seconds — it settles quickly |
        | Saga parked, caller is a customer | `409` | Contact support; a human is already on it |
        | Saga parked, caller is an administrator | `200` | Allowed, but verify the payment first |
        | Already cancelled | `409` | Nothing to do |

        The running-saga refusal is the load-bearing one. A mid-flight order has a command
        outstanding, and cancelling into that race can issue a refund for a charge the system has
        not finished recording.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        description: Optional. The reason is recorded and shown to the customer.
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 255
                  example: Ordered the wrong size
      responses:
        "200":
          description: The cancelled order. Undoing continues asynchronously.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/orders/number/{orderNumber}:
    get:
      tags: [Orders]
      summary: Fetch one order by its human readable number
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema: { type: string, example: CF-20260826-000123 }
      responses:
        "200":
          description: The order
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ------------------------------------------------------------------ basket
...
  /api/payments/order/{orderId}:
    get:
      tags: [Payments]
      summary: Fetch the payment for an order
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The payment
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Payment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ------------------------------------------------------------------ notifications
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token returned by `POST /api/auth/login`

  # This file is the hand-written contract for the paths that matter most. It is deliberately
  # not exhaustive: each service serves its own generated definition at /v3/api-docs, aggregated
  # behind the gateway at /swagger-ui.html, and that is authoritative for everything else --
  # notably the back-office endpoints for warehouses, stock levels, product images and review
  # moderation, which no external client calls.
  responses:
    Unauthorized:
      description: No valid access token (`UNAUTHORIZED`, `TOKEN_EXPIRED`, `TOKEN_INVALID`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationFailed:
      description: Request validation failed; `fieldErrors` lists every violated constraint
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: |
        The request is well formed but the resource is not in a state that allows it
        (`ORDER_NOT_MODIFIABLE`, `CONFLICT`). The message says what to do about it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ApiResponse:
      type: object
      description: Uniform success envelope.
      properties:
        success: { type: boolean, example: true }
        data: { description: The payload }
        message: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        correlationId:
          type: string
          nullable: true
          description: Quote this in a support request; it stitches the saga together in the logs.
    ErrorResponse:
      type: object
      description: Uniform error envelope.
      properties:
        success: { type: boolean, example: false }
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 409 }
        error: { type: string, example: Conflict }
        code:
          type: string
          example: INSUFFICIENT_STOCK
          description: Machine readable error code; stable across releases.
        message: { type: string }
        path: { type: string, example: /api/orders }
        correlationId: { type: string, nullable: true }
        fieldErrors:
          type: array
          nullable: true
          items:
            type: object
            properties:
              field: { type: string, example: email }
              rejectedValue: {}
              message: { type: string, example: must be a valid address }
    Page:
      type: object
      properties:
        content: { type: array, items: {} }
        page: { type: integer, example: 0 }
        size: { type: integer, example: 20 }
        totalElements: { type: integer, format: int64, example: 137 }
        totalPages: { type: integer, example: 7 }
        first: { type: boolean }
        last: { type: boolean }
...
    CreateOrderRequest:
      type: object
      required: [items, shippingAddress]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId: { type: string, format: uuid }
              quantity: { type: integer, minimum: 1, maximum: 999 }
        shippingAddress:
          type: string
          maxLength: 500
          example: Keizersgracht 1, 1015 CJ Amsterdam, NL
        idempotencyKey:
          type: string
          maxLength: 64
          nullable: true
          description: Replaying this key returns the original order instead of creating a second one.
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderNumber: { type: string, example: CF-20260826-000123 }
        userId: { type: string, format: uuid }
        userEmail: { type: string, format: email }
        status:
          type: string
          enum: [CREATED, INVENTORY_RESERVED, PAID, COMPLETED, CANCELLED]
        totalAmount: { type: number, format: double, example: 3798.00 }
        currency: { type: string, example: EUR }
        shippingAddress: { type: string }
        itemCount: { type: integer, example: 3 }
        items:
          type: array
          items:
            type: object
            properties:
              productId: { type: string, format: uuid }
              sku: { type: string }
              productName:
                type: string
                description: Snapshotted at order time; a later catalogue change never rewrites it.
              quantity: { type: integer }
              unitPrice: { type: number, format: double }
              subtotal: { type: number, format: double }
        paymentId: { type: string, format: uuid, nullable: true }
        failedStep:
          type: string
          nullable: true
          enum: [INVENTORY, PAYMENT]
          description: Present when the order was cancelled.
        failureReason: { type: string, nullable: true, example: OUT_OF_STOCK }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
    Payment:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        orderNumber: { type: string, nullable: true }
        userId: { type: string, format: uuid, nullable: true }
        amount: { type: number, format: double }
        currency: { type: string, example: EUR }
        status: { type: string, enum: [PENDING, COMPLETED, FAILED, REFUNDED] }
        method: { type: string, enum: [CARD, IDEAL, PAYPAL, BANK_TRANSFER] }
        transactionId:
          type: string
          nullable: true
          description: Acquirer reference; present once the charge is approved.
        failureReason: { type: string, nullable: true, example: INSUFFICIENT_FUNDS }
        createdAt: { type: string, format: date-time }
        processedAt: { type: string, format: date-time, nullable: true }
...